refactor(studio): split routes/index into per-domain route groups
Break the ~770-line routes/index.tsx into one feature-gated route module
per domain under routes/groups/, leaving index.tsx as a thin composition
root that spreads the groups into the workspace route tree.

Each group owns its own lazy component definitions and exports a fully
gated RouteObject[]. Hoist the lazy-load Suspense boundary to the single
wrapper around the workspace Outlet rather than repeating it per route.
