+++ web/packages/studio/src/routes/groups/index.ts
@@ -1,0 +1,21 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { dashboardRoutes } from '@studio/routes/groups/dashboardRoutes';
+export { baseModelsRoutes } from '@studio/routes/groups/baseModelsRoutes';
+export { filesetRoutes } from '@studio/routes/groups/filesetRoutes';
+export { secretsRoutes } from '@studio/routes/groups/secretsRoutes';
+export { guardrailsRoutes } from '@studio/routes/groups/guardrailsRoutes';
+export { inferenceProviderRoutes } from '@studio/routes/groups/inferenceProviderRoutes';
+export { deploymentRoutes } from '@studio/routes/groups/deploymentRoutes';
+export { evaluationRoutes } from '@studio/routes/groups/evaluationRoutes';
+export { experimentRoutes } from '@studio/routes/groups/experimentRoutes';
+export { customizationRoutes } from '@studio/routes/groups/customizationRoutes';
+export { jobRoutes } from '@studio/routes/groups/jobRoutes';
+export { intakeRoutes } from '@studio/routes/groups/intakeRoutes';
+export { safeSynthesizerRoutes } from '@studio/routes/groups/safeSynthesizerRoutes';
+export { dataDesignerRoutes } from '@studio/routes/groups/dataDesignerRoutes';
+export { agentRoutes } from '@studio/routes/groups/agentRoutes';
+export { settingsRoutes } from '@studio/routes/groups/settingsRoutes';
+export { modelCompareRoutes } from '@studio/routes/groups/modelCompareRoutes';
+export { memberRoutes } from '@studio/routes/groups/memberRoutes';
